Changeds RegisterPlantModal.tsx to CreatePlantModal.tsx
