+++ src/front-end/src/routes/_dashboard/_dashboard.plants.tsx
@@ -1,7 +1,7 @@
 import { createFileRoute } from '@tanstack/react-router'
 import { plantsMock } from '../admin/plants'
 import PlantCard from '@/components/PlantCard'
-import RegisterPlantModal from '@/components/RegisterPlantModal'
+import CreatePlantModal from '@/components/CreatePlantModal'
 
 export const Route = createFileRoute('/_dashboard/_dashboard/plants')({
   component: RouteComponent,
@@ -18,7 +18,7 @@ function RouteComponent() {
           <h1 className='font-bold text-2xl mb-2'>My Plants</h1>
           <p className='text-gray-600 text-sm'>Manage and monitor your plant collection</p>
         </div>
-        <RegisterPlantModal />
+        <CreatePlantModal />
       </div>
 
       <div className='grid gap-4 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4'>
